moved element controls to gui.controls
