+++ ecris/csd/viewer/gui/__init__.py
@@ -1,6 +1,5 @@
-from .controls import PlotControls, FileListControls, FileList
+from .controls import PlotControls, FileListControls, FileList, ElementButtons
 from .plot import Plot
-from .elements import ElementButtons
 from .menu import AppMenu
 from .diagnostic_window import DiagnosticWindow
 from .file_info_pane import FileInfoPane
