+++ ecris/csd/viewer/gui/controls/__init__.py
@@ -1,2 +1,3 @@
 from .controls import FileListControls, PlotControls
+from .elements import ElementButtons
 from .file_list import FileList
